+++ tests/tests_unit/test_api/test_events.py
@@ -56,6 +56,34 @@ class TestEvents:
         EVENTS_API.list(partitions=13, limit=float("inf"))
         assert 13 == len(mock_events_response.calls)
 
+    def test_call_root(self, mock_events_response):
+        list(EVENTS_API.__call__(root_asset_ids=[23], root_asset_external_ids=["a", "b"], limit=10))
+        calls = mock_events_response.calls
+        assert 1 == len(calls)
+        assert {
+            "cursor": None,
+            "limit": 10,
+            "filter": {"rootAssetIds": [{"id": 23}, {"externalId": "a"}, {"externalId": "b"}]},
+        } == jsgz_load(calls[0].request.body)
+
+    def test_list_root_ids_list(self, mock_events_response):
+        EVENTS_API.list(root_asset_ids=[1, 2], limit=10)
+        calls = mock_events_response.calls
+        assert 1 == len(calls)
+        assert {"cursor": None, "limit": 10, "filter": {"rootAssetIds": [{"id": 1}, {"id": 2}]}} == jsgz_load(
+            calls[0].request.body
+        )
+
+    def test_list_root_extids_list(self, mock_events_response):
+        EVENTS_API.list(root_asset_external_ids=["1", "2"], limit=10)
+        calls = mock_events_response.calls
+        assert 1 == len(calls)
+        assert {
+            "cursor": None,
+            "limit": 10,
+            "filter": {"rootAssetIds": [{"externalId": "1"}, {"externalId": "2"}]},
+        } == jsgz_load(calls[0].request.body)
+
     def test_create_single(self, mock_events_response):
         res = EVENTS_API.create(Event(external_id="1"))
         assert isinstance(res, Event)